refactor table drawing to tables, bug fix testing
