+++ src/clusternetwork.js
@@ -1,6 +1,5 @@
 import * as d3 from "d3";
 import _ from "underscore";
-import jsConvert from "js-convert-case";
 import * as topojson from "topojson";
 import * as colorPicker from "./colorPicker.js";
 import * as scatterPlot from "./scatterplot.js";
@@ -11,14 +10,14 @@ import * as clustersOfInterest from "./clustersOI/clusterOI.js";
 import { hivtrace_cluster_depthwise_traversal } from "./helpers.js";
 import * as svgPlots from "./svgPlots.js";
 
-const _networkSubclusterSeparator = ".";
+export const _networkSubclusterSeparator = ".";
 export const _networkGraphAttrbuteID = "patient_attribute_schema";
-const _networkNodeAttributeID = "patient_attributes";
+export const _networkNodeAttributeID = "patient_attributes";
 export const _networkMissing = __("general")["missing"];
-const _networkReducedValue = "Different (other) value";
-const _networkMissingOpacity = "0.1";
+export const _networkReducedValue = "Different (other) value";
+export const _networkMissingOpacity = "0.1";
 export const _networkMissingColor = "#999";
-const _networkContinuousColorStops = 9;
+export const _networkContinuousColorStops = 9;
 export const _networkWarnExecutiveMode =
   "This feature is not available in the executive mode.";
 
@@ -33,7 +32,7 @@ const _networkShapeOrdering = [
   "pentagon",
 ];
 
-const _defaultFloatFormat = d3.format(",.2r");
+export const _defaultFloatFormat = d3.format(",.2r");
 export const _defaultPercentFormat = d3.format(",.3p");
 export const _defaultPercentFormatShort = d3.format(".2p");
 export const _defaultDateFormats = [d3.time.format.iso];
@@ -42,7 +41,6 @@ export const _defaultDateViewFormatMMDDYYY = d3.time.format("%m%d%Y");
 export const _defaultDateViewFormat = d3.time.format("%b %d, %Y");
 const _defaultDateViewFormatShort = d3.time.format("%B %Y");
 export const _defaultDateViewFormatSlider = d3.time.format("%Y-%m-%d");
-const _networkDotFormatPadder = d3.format("08d");
 export const _defaultDateViewFormatClusterCreate = d3.time.format("%Y%m");
 
 const _networkCategoricalBase = [
@@ -922,7 +920,7 @@ var hivtrace_cluster_network_graph = function (
         },
         generator: function (node) {
           return {
-            callback: function (element, payload) {
+            callback: function (self, element, payload) {
               //payload = _.filter (payload, function (d) {return d});
               var this_cell = d3.select(element);
 
@@ -1007,7 +1005,7 @@ var hivtrace_cluster_network_graph = function (
       (not_nested ? "" : "#" + button_bar_ui) + helpers.get_ui_element_selector_by_role(role)
     );
   };
-  
+
   //---------------------------------------------------------------------------------------------------
   // BEGIN: NODE SET EDITOR
   //---------------------------------------------------------------------------------------------------
@@ -2897,7 +2895,7 @@ var hivtrace_cluster_network_graph = function (
       _.each(list, set_init_coords);
     });
 
-    self.clusters.forEach(collapse_cluster);
+    self.clusters.forEach(self.collapse_cluster);
   }
 
   function change_spacing(delta) {
@@ -3093,6 +3091,9 @@ var hivtrace_cluster_network_graph = function (
     self.exclude_cluster_ids = {};
     self.has_hxb2_links = false;
     self.cluster_sizes = [];
+    self.update_volatile_elements = (container) => {
+      tables.update_volatile_elements(self, container);
+    };
 
     graph_data.Nodes.forEach((d) => {
       if (typeof self.cluster_sizes[d.cluster - 1] === "undefined") {
@@ -3399,6 +3400,7 @@ var hivtrace_cluster_network_graph = function (
           });
 
           tables.add_a_sortable_table(
+            self,
             modal.select(
               self.get_ui_element_selector_by_role(
                 "overlap_list_data_table",
@@ -3533,7 +3535,7 @@ var hivtrace_cluster_network_graph = function (
             rows.push([
               {
                 value: pg,
-                callback: function (element, payload) {
+                callback: function (self, element, payload) {
                   var this_cell = d3.select(element);
                   this_cell
                     .append("input")
@@ -3555,6 +3557,7 @@ var hivtrace_cluster_network_graph = function (
           });
 
           tables.add_a_sortable_table(
+            self,
             modal.select(
               self.get_ui_element_selector_by_role(
                 "priority_set_merge_table",
@@ -4628,208 +4631,6 @@ var hivtrace_cluster_network_graph = function (
     self.update();
   }
 
-  function _cluster_table_draw_id(element, payload) {
-    var this_cell = d3.select(element);
-    this_cell.selectAll("*").remove();
-    const _is_subcluster = payload[1];
-    var cluster_id = payload[0];
-
-    if (_is_subcluster) {
-      //console.log (payload);
-
-      //this_cell.append("i")
-      //      .classed("fa fa-arrow-circle-o-right", true).style("padding-right", "0.25em");
-
-      /*if (payload[2].rr_count) {
-        this_cell
-          .append("i")
-          .classed("fa fa-exclamation-triangle", true)
-          .attr("title", "Subcluster has recent/rapid nodes");
-      }*/
-      this_cell.append("span").text(cluster_id).style("padding-right", "0.5em");
-
-      this_cell
-        .append("button")
-        .classed("btn btn-sm pull-right", true)
-        //.text(__("clusters_tab")["view"])
-        .on("click", (e) => {
-          self.view_subcluster(payload[2]);
-        })
-        .append("i")
-        .classed("fa fa-eye", true)
-        .attr("title", __("clusters_tab")["view"]);
-    } else {
-      this_cell.append("span").text(cluster_id).style("padding-right", "0.5em");
-      this_cell
-        .append("button")
-        .classed("btn btn-sm pull-right", true)
-        .style("margin-right", "0.25em")
-        .on("click", (e) => {
-          self.open_exclusive_tab_view(cluster_id);
-        })
-        .append("i")
-        .classed("fa fa-eye", true)
-        .attr("title", __("clusters_tab")["view"]);
-    }
-    this_cell
-      .append("button")
-      .classed("btn btn-sm pull-right", true)
-      .style("margin-right", "0.25em")
-      //.text(__("clusters_tab")["list"])
-      .attr("data-toggle", "modal")
-      .attr(
-        "data-target",
-        self.get_ui_element_selector_by_role("cluster_list", true)
-      )
-      .attr("data-cluster", cluster_id)
-      .append("i")
-      .classed("fa fa-list", true)
-      .attr("title", __("clusters_tab")["list"]);
-  }
-
-  function _cluster_table_draw_buttons(element, payload) {
-    var this_cell = d3.select(element);
-    const label_diff = function (c_info) {
-      const d = c_info["delta"];
-      const moved = c_info["moved"];
-      const deleted = c_info["deleted"];
-      const new_count = c_info["new_nodes"] ? c_info["new_nodes"] : 0;
-
-      /*if (moved) {
-            if (d > 0) {
-                return "" + moved + " nodes moved +" + d + " new";
-            } else {
-                if (d === 0) {
-                    return "" + moved + " nodes moved";
-                } else {
-                    return "" + moved + " nodes moved " + (-d) + " removed";
-                }
-            }
-
-        } else {
-            if (d > 0) {
-                return "+" + d + " nodes";
-            } else {
-                if (d === 0) {
-                    return "no size change";
-                } else {
-                    return "" + (-d) + " nodes removed";
-                }
-            }
-        }*/
-
-      let label_str = "";
-      if (moved) label_str = " " + moved + " moved ";
-      if (new_count) label_str += "+" + new_count + " new ";
-      if (deleted) label_str += "-" + deleted + " previous ";
-      return label_str;
-    };
-
-    var labels = [];
-
-    if (payload[4]) {
-      if (payload[4]["type"] === "new") {
-        if (payload[4]["moved"]) {
-          labels.push(["renamed " + label_diff(payload[4]), 2]);
-        } else {
-          labels.push(["new", 3]);
-        }
-      } else if (payload[4]["type"] === "extended") {
-        labels.push([label_diff(payload[4]), payload["4"]["flag"]]);
-      } else if (payload[4]["type"] === "merged") {
-        labels.push([
-          "Merged " +
-          payload[4]["old_clusters"].join(", ") +
-          " " +
-          label_diff(payload[4]),
-          payload["4"]["flag"],
-        ]);
-      }
-    }
-
-    labels.push([
-      [
-        payload[0]
-          ? __("clusters_tab")["expand"]
-          : __("clusters_tab")["collapse"],
-        payload[0] ? "fa-expand" : "fa-compress",
-      ],
-      0,
-    ]);
-    if (payload[1]) {
-      labels.push([["problematic", "fa-exclamation-circle"], 1]);
-    }
-    if (payload[2]) {
-      labels.push([["match", "fa-check-square"], 1]);
-    }
-    var buttons = this_cell.selectAll("button").data(labels);
-    buttons.enter().append("button");
-    buttons.exit().remove();
-    buttons
-      .classed("btn btn-xs", true)
-      .classed("btn-default", (d) => d[1] !== 1 && d[1] !== 2)
-      .classed("btn-danger", (d) => d[1] === 2)
-      .classed("btn-success", (d) => d[1] === 3)
-      /*.text(function(d) {
-        return d[0];
-      })*/
-      .style("margin-right", "0.25em")
-      .attr("disabled", (d) => d[1] === 1 ? "disabled" : null)
-      .on("click", (d) => {
-        if (d[1] === 0) {
-          if (payload[0]) {
-            expand_cluster(self.clusters[payload[3] - 1], true);
-          } else {
-            collapse_cluster(self.clusters[payload[3] - 1]);
-          }
-          tables.update_volatile_elements(self.cluster_table);
-          if (self.subcluster_table) {
-            tables.update_volatile_elements(self.subcluster_table);
-          }
-        } else if (d[1] === 2 || d[1] === 3) {
-          //_social_view_options (labeled_links, shown_types),
-
-          var shown_types = { Existing: 1, "Newly added": 1 },
-            link_class = ["Existing", "Newly added"];
-
-          self
-            .open_exclusive_tab_view(
-              payload[3],
-              null,
-              (cluster_id) => "Cluster " + cluster_id + " [changes view]",
-              self._social_view_options(link_class, shown_types, (e) => {
-                if (_.isObject(e.source) && self._is_new_node(e.source))
-                  return "Newly added";
-                if (_.isObject(e.target) && self._is_new_node(e.target))
-                  return "Newly added";
-
-                return e.attributes.indexOf("added-to-prior") >= 0
-                  ? "Newly added"
-                  : "Existing";
-              })
-            )
-            .handle_attribute_categorical("_newly_added");
-        }
-      });
-    buttons.each(function (d, i) {
-      var this_e = d3.select(this);
-      if (_.isString(d[0])) {
-        this_e.selectAll("i").remove();
-        this_e.text(d[0]);
-      } else {
-        var i_span = this_e.selectAll("i").data([d[0]]);
-        i_span.enter().append("i");
-        i_span
-          .attr(
-            "class",
-            (d) => "fa " + d[1],
-            true
-          )
-          .attr("title", (d) => d[0]);
-      }
-    });
-  }
-
   self._extract_single_cluster = function (
     nodes,
     filter,
@@ -4889,637 +4690,6 @@ var hivtrace_cluster_network_graph = function (
     return cluster_json;
   }
 
-  function _node_table_draw_buttons(element, payload) {
-    var this_cell = d3.select(element);
-    let labels;
-    if (payload.length === 1) {
-      if (_.isString(payload[0])) {
-        labels = [[payload[0], 1, "btn-warning"]];
-      } else {
-        labels = ["can't be shown", 1];
-      }
-    } else {
-      labels = [[payload[0] ? "hide" : "show", 0]];
-      // TODO: deprecated? remove if not needed (5/22/2024 meeting with @spond, @daniel-ji, @stevenweaver)
-    }
-
-    if (payload.length === 2 && payload[1] >= 1) {
-      labels.push([
-        "view cluster",
-        function () {
-          self.open_exclusive_tab_view(payload[1]);
-        },
-      ]);
-    }
-
-    var buttons = this_cell.selectAll("button").data(labels);
-    buttons.enter().append("button");
-    buttons.exit().remove();
-    buttons
-      .classed("btn btn-xs btn-node-property", true)
-      .classed("btn-primary", true)
-      //.classed(function (d) {return d.length >=3 ? d[2] : "";}, function (d) {return d.length >= 3;})
-      .text((d) => d[0])
-      .attr("disabled", (d) => d[1] && !_.isFunction(d[1]) ? "disabled" : null)
-      .on("click", (d) => {
-        if (_.isFunction(d[1])) {
-          d[1].call(d);
-        } else if (d[1] === 0) {
-          if (payload[0]) {
-            collapse_cluster(self.clusters[payload[3] - 1], true);
-          } else {
-            expand_cluster(self.clusters[payload[3] - 1]);
-          }
-          //tables.format_a_cell(d3.select(element).datum(), null, element);
-          tables.update_volatile_elements(nodesTab.getNodeTable());
-        }
-      });
-    buttons.each(function (d, e) {
-      if (d.length >= 3) {
-        d3.select(this).classed("btn-primary", false).classed(d[2], true);
-      }
-    });
-  }
-
-  /*self.process_table_volatile_event = function (e) {
-    console.log (e);
-    e.detail
-      .selectAll("td")
-      .filter(function(d) {
-        return "volatile" in d;
-      })
-      .each(function(d, i) {
-        tables.format_a_cell(d, i, this);
-      });
-  };*/
-
-  self.update_volatile_elements = tables.update_volatile_elements;
-
-  self.draw_extended_node_table = function (
-    node_list,
-    container,
-    extra_columns
-  ) {
-    container = container || nodesTab.getNodeTable();
-
-    if (container) {
-      node_list = node_list || self.nodes;
-      var column_ids = self._extract_exportable_attributes(true);
-
-      self.displayed_node_subset = _.filter(
-        _.map(self.displayed_node_subset, (n, i) => {
-          if (_.isString(n)) {
-            n = _.find(column_ids, (cd) => cd.raw_attribute_key === n);
-
-            if (n) {
-              return n;
-            }
-            return column_ids[i];
-          }
-          return n;
-        }),
-        (c) => c
-      );
-
-      var node_data = self._extract_attributes_for_nodes(
-        node_list,
-        self.displayed_node_subset
-      );
-      node_data.splice(0, 1);
-      var table_headers = _.map(
-        self.displayed_node_subset,
-        (n, col_id) => ({
-          value: n.raw_attribute_key,
-          sort: "value",
-          filter: true,
-          volatile: true,
-          help: "label" in n ? n.label : n.raw_attribute_key,
-          //format: (d) => "label" in d ? d.label : d.raw_attribute_key,
-          callback: function (element, payload) {
-            var dropdown = d3
-              .select(element)
-              .append("div")
-              .classed("dropdown", true);
-            // add col_id to ensure that the dropdowns are unique
-            var menu_id = "hivtrace_node_column_" + payload + "_" + col_id;
-            var dropdown_button = dropdown
-              .append("button")
-              .classed({
-                btn: true,
-                "btn-default": true,
-                "btn-xs": true,
-                "dropdown-toggle": true,
-              })
-              .attr("type", "button")
-              .attr("data-toggle", "dropdown")
-              .attr("aria-haspopup", "true")
-              .attr("aria-expanded", "false")
-              .attr("id", menu_id);
-
-            function format_key(key) {
-              const formattedKey = jsConvert.toHeaderCase(key);
-              const words = formattedKey.split(" ");
-              const mappedWords = _.map(words, (word) => {
-                if (word.toLowerCase() === "hivtrace") {
-                  return "HIV-TRACE";
-                }
-                if (word.toLowerCase() === "id") {
-                  return "ID";
-                }
-
-                return word;
-              });
-              return mappedWords.join(" ");
-            }
-
-            function get_text_label(key) {
-              return key in json.patient_attribute_schema
-                ? json.patient_attribute_schema[key].label
-                : format_key(key);
-            }
-
-            dropdown_button.text(get_text_label(payload));
-
-            dropdown_button.append("i").classed({
-              fa: true,
-              "fa-caret-down": true,
-              "fa-lg": true,
-            });
-            var dropdown_list = dropdown
-              .append("ul")
-              .classed("dropdown-menu", true)
-              .attr("aria-labelledby", menu_id);
-
-            dropdown_list = dropdown_list.selectAll("li").data(
-              _.filter(column_ids, (alt) => alt.raw_attribute_key !== n.raw_attribute_key)
-            );
-            dropdown_list.enter().append("li");
-            dropdown_list.each(function (data, i) {
-              var handle_change = d3
-                .select(this)
-                .append("a")
-                .attr("href", "#")
-                .text((data) => get_text_label(data.raw_attribute_key));
-              handle_change.on("click", (d) => {
-                self.displayed_node_subset[col_id] = d;
-                self.draw_extended_node_table(
-                  node_list,
-                  container,
-                  extra_columns
-                );
-              });
-            });
-            return dropdown;
-          },
-        })
-      );
-
-      if (extra_columns) {
-        _.each(extra_columns, (d) => {
-          if (d.prepend) {
-            table_headers.splice(0, 0, d.description);
-          } else {
-            table_headers.push(d.description);
-          }
-        });
-      }
-      //console.log (self.displayed_node_subset);
-
-      var table_rows = node_data.map((n, i) => {
-        var this_row = _.map(n, (cell, c) => {
-          let cell_definition = null;
-
-          if (self.displayed_node_subset[c].type === "Date") {
-            cell_definition = {
-              value: cell,
-              format: function (v) {
-                if (v === _networkMissing) {
-                  return v;
-                }
-                return _defaultDateViewFormatSlider(v);
-              },
-            };
-          } else if (self.displayed_node_subset[c].type === "Number") {
-            cell_definition = { value: cell, format: d3.format(".2f") };
-          }
-          if (!cell_definition) {
-            cell_definition = { value: cell };
-          }
-
-          // this makes the table rendering too slow
-
-          /*if (c === 0 && self._is_CDC_) {
-             cell_definition.volatile = true;
-             cell_definition.actions = function (item, value) {
-              if (!clustersOfInterest.get_editor()) {
-                    return null;
-              } else {
-                    return [
-                        {
-                            "icon"   : "fa-plus-square",
-                            "action" : function (button,v) {
-                                if (clustersOfInterest.get_editor()) {
-                                    clustersOfInterest.get_editor().append_node_objects (d.children);
-                                }
-                                return false;
-                            },
-                            "help"   : "Add to priority set"
-                        }
-                    ];
-                }
-            };
-          }*/
-
-          return cell_definition;
-        });
-
-        if (extra_columns) {
-          _.each(extra_columns, (ed) => {
-            if (ed.prepend) {
-              this_row.splice(0, 0, ed.generator(node_list[i], self));
-            } else {
-              this_row.push(ed.generator(node_list[i], self));
-            }
-          });
-        }
-
-        return this_row;
-      });
-
-      self.draw_node_table(
-        null,
-        null,
-        [table_headers],
-        table_rows,
-        container,
-        'Showing <span class="badge" data-hivtrace-ui-role="table-count-shown">--</span>/<span class="badge" data-hivtrace-ui-role="table-count-total">--</span> network nodes'
-      );
-    }
-  };
-
-  self.draw_node_table = function (
-    extra_columns,
-    node_list,
-    headers,
-    rows,
-    container,
-    table_caption
-  ) {
-    container = container || nodesTab.getNodeTable();
-
-    if (container) {
-      node_list = node_list || self.nodes;
-
-      if (!headers) {
-        headers = [
-          [
-            {
-              value: "ID",
-              sort: "value",
-              help: "Node ID",
-            },
-            {
-              value: "Action",
-              sort: "value",
-            },
-            {
-              value: "# of links",
-              sort: "value",
-              help: "Number of links (Node degree)",
-            },
-            {
-              value: "Cluster",
-              sort: "value",
-              help: "Which cluster does the node belong to",
-            },
-          ],
-        ];
-
-        if (extra_columns) {
-          _.each(extra_columns, (d) => {
-            if (d.prepend) {
-              headers[0].splice(0, 0, d.description);
-            } else {
-              headers[0].push(d.description);
-            }
-          });
-        }
-
-        rows = node_list.map((n, i) => {
-          var this_row = [
-            {
-              value: n.id,
-              help: "Node ID",
-            },
-            {
-              value: function () {
-                if (n.node_class !== "injected") {
-                  try {
-                    if (self.exclude_cluster_ids[n.cluster]) {
-                      // parent cluster can't be rendered
-                      // because of size restrictions
-                      return [n.cluster];
-                    }
-                    return [
-                      !self.clusters[self.cluster_mapping[n.cluster]].collapsed,
-                      n.cluster,
-                    ];
-                  } catch (err) {
-                    return [-1];
-                  }
-                } else {
-                  return [n.node_annotation];
-                }
-              },
-              callback: _node_table_draw_buttons,
-              volatile: true,
-            },
-            {
-              value: "degree" in n ? n.degree : "Not defined",
-              help: "Node degree",
-            },
-            {
-              value: "cluster" in n ? n.cluster : "Not defined",
-              help: "Which cluster does the node belong to",
-            },
-          ];
-
-          if (extra_columns) {
-            _.each(extra_columns, (ed) => {
-              if (ed.prepend) {
-                this_row.splice(0, 0, ed.generator(n, self));
-              } else {
-                this_row.push(ed.generator(n, self));
-              }
-            });
-          }
-          return this_row;
-        });
-      }
-
-      tables.add_a_sortable_table(
-        container,
-        headers,
-        rows,
-        true,
-        table_caption,
-        clustersOfInterest.get_editor()
-        // rows
-      );
-    }
-  };
-
-  self.draw_cluster_table = function (extra_columns, element, options) {
-    var skip_clusters = options && options["no-clusters"];
-    var skip_subclusters = !(options && options["subclusters"]);
-
-    element = element || self.cluster_table;
-
-    if (element) {
-      var headers = [
-        [
-          {
-            value: __("general")["cluster"] + " ID",
-            sort: function (c) {
-              return _.map(
-                c.value[0].split(_networkSubclusterSeparator),
-                (ss) => _networkDotFormatPadder(Number(ss))
-              ).join("|");
-            },
-            help: "Unique cluster ID",
-          },
-          {
-            value: __("general")["attributes"],
-            sort: function (c) {
-              c = c.value();
-              if (c[4]) {
-                // has attributes
-                return c[4]["delta"];
-              }
-              return c[0];
-            },
-            help: "Visibility in the network tab and other attributes",
-          },
-          {
-            value: __("clusters_tab")["size"],
-            sort: "value",
-            help: "Number of nodes in the cluster",
-          },
-        ],
-      ];
-
-      if (self.cluster_attributes) {
-        headers[0][1]["presort"] = "desc";
-      }
-
-      if (self._is_seguro) {
-        headers[0].push({
-          value: __("clusters_tab")["number_of_genotypes_in_past_2_months"],
-          sort: "value",
-          help: "# of cases in cluster genotyped in the last 2 months",
-        });
-
-        headers[0].push({
-          value:
-            __("clusters_tab")["scaled_number_of_genotypes_in_past_2_months"],
-          sort: "value",
-          help: "# of cases in cluster genotyped in the last 2 months divided by the square-root of the cluster size",
-        });
-      }
-
-      if (!self._is_CDC_) {
-        headers[0].push({
-          value:
-            __("statistics")["links_per_node"] +
-            "<br>" +
-            __("statistics")["mean"] +
-            "[" +
-            __("statistics")["median"] +
-            ", IQR]",
-          html: true,
-        });
-
-        headers[0].push({
-          value:
-            __("statistics")["genetic_distances_among_linked_nodes"] +
-            "<br>" +
-            __("statistics")["mean"] +
-            "[" +
-            __("statistics")["median"] +
-            ", IQR]",
-          help: "Genetic distance among nodes in the cluster",
-          html: true,
-        });
-      }
-
-      if (extra_columns) {
-        _.each(extra_columns, (d) => {
-          headers[0].push(d.description);
-        });
-      }
-
-      if (options && options["headers"]) {
-        options["headers"](headers);
-      }
-
-      var rows = [];
-
-      _.each(self.clusters, (cluster) => {
-        var make_row = function (d, is_subcluster) {
-          var this_row = [
-            {
-              value: [d.cluster_id, is_subcluster, d], //.cluster_id,
-              callback: _cluster_table_draw_id,
-            },
-            {
-              value: function () {
-                var actual_cluster = is_subcluster ? d.parent_cluster : d;
-
-                return [
-                  actual_cluster.collapsed,
-                  actual_cluster.hxb2_linked,
-                  actual_cluster.match_filter,
-                  actual_cluster.cluster_id,
-                  is_subcluster
-                    ? null
-                    : self.cluster_attributes
-                      ? self.cluster_attributes[actual_cluster.cluster_id]
-                      : null,
-                ];
-              },
-              callback: _cluster_table_draw_buttons,
-              volatile: true,
-            },
-            {
-              value: d.children.length,
-            },
-          ];
-
-          if (self._is_CDC_) {
-            this_row[2].volatile = true;
-            this_row[2].actions = function (item, value) {
-              if (!clustersOfInterest.get_editor()) {
-                return null;
-              }
-              return [
-                {
-                  icon: "fa-plus",
-                  action: function (button, v) {
-                    if (clustersOfInterest.get_editor()) {
-                      clustersOfInterest.get_editor().append_node_objects(
-                        d.children
-                      );
-                    }
-                    return false;
-                  },
-                  help: "Add to cluster of interest",
-                },
-              ];
-            };
-          }
-
-          if (self._is_seguro) {
-            this_row.push({
-              value: d,
-              format: function (d) {
-                return _.filter(
-                  d.children,
-                  (child) =>
-                    d3.time.months(
-                      child.patient_attributes["sample_dt"],
-                      helpers.getCurrentDate()
-                    ).length <= 2
-                ).length;
-              },
-            });
-
-            this_row.push({
-              value: d,
-              format: function (d) {
-                const recent = _.filter(
-                  d.children,
-                  (child) =>
-                    d3.time.months(
-                      child.patient_attributes["sample_dt"],
-                      helpers.getCurrentDate()
-                    ).length <= 2
-                ).length;
-                return recent / Math.sqrt(d.children.length);
-              },
-            });
-          }
-
-          if (!self._is_CDC_) {
-            this_row.push({
-              value: d.degrees,
-              format: function (d) {
-                try {
-                  return (
-                    _defaultFloatFormat(d["mean"]) +
-                    " [" +
-                    _defaultFloatFormat(d["median"]) +
-                    ", " +
-                    _defaultFloatFormat(d["Q1"]) +
-                    " - " +
-                    _defaultFloatFormat(d["Q3"]) +
-                    "]"
-                  );
-                } catch (e) {
-                  return "";
-                }
-              },
-            });
-            this_row.push({
-              value: d.distances,
-              format: function (d) {
-                try {
-                  return (
-                    _defaultFloatFormat(d["mean"]) +
-                    " [" +
-                    _defaultFloatFormat(d["median"]) +
-                    ", " +
-                    _defaultFloatFormat(d["Q1"]) +
-                    " - " +
-                    _defaultFloatFormat(d["Q3"]) +
-                    "]"
-                  );
-                } catch (e) {
-                  return "";
-                }
-              },
-            });
-          }
-          if (extra_columns) {
-            _.each(extra_columns, (ed) => {
-              this_row.push(ed.generator(d, self));
-            });
-          }
-
-          return this_row;
-        };
-
-        if (!skip_clusters) {
-          rows.push(make_row(cluster, false));
-        }
-
-        if (!skip_subclusters) {
-          _.each(cluster.subclusters, (sub_cluster) => {
-            rows.push(make_row(sub_cluster, true));
-          });
-        }
-      });
-
-      tables.add_a_sortable_table(
-        element,
-        headers,
-        rows,
-        true,
-        options && options["caption"] ? options["caption"] : null,
-        clustersOfInterest.get_editor()
-      );
-    }
-  };
-
   /*------------ Update layout code ---------------*/
   function update_network_string(node_count, edge_count) {
     if (network_status_string) {
@@ -5535,7 +4705,7 @@ var hivtrace_cluster_network_graph = function (
       // const nodes_removed = graph_data.Nodes.length - singletons - self.nodes.length;
       // const networkString = "Displaying a network on <strong>" + self.nodes.length + "</strong> nodes, <strong>" + self.clusters.length + "</strong> clusters"
       //         + (clusters_removed > 0 ? " (an additional " + clusters_removed + " clusters and " + nodes_removed + " nodes have been removed due to network size constraints)" : "") + ". <strong>"
-      //         + clusters_shown +"</strong> clusters are expanded. Of <strong>" + self.edges.length + "</strong> edges, <strong>" + draw_me.edges.length + "</strong>, and of  <strong>" + self.nodes.length  + " </strong> nodes,  <strong>" + draw_me.nodes.length + " </strong> are displayed. ";
+      //         + clusters_shown +"</strong> clusters are expanded. Of <strong>" + self.edges.length + "</strong> edges, <strong>" + self.draw_me.edges.length + "</strong>, and of  <strong>" + self.nodes.length  + " </strong> nodes,  <strong>" + self.draw_me.nodes.length + " </strong> are displayed. ";
       // if (singletons > 0) {
       //   networkString += "<strong>" +singletons + "</strong> singleton nodes are not shown. ";
       // }
@@ -6822,14 +5992,14 @@ var hivtrace_cluster_network_graph = function (
     var rendered_nodes, rendered_clusters, link;
 
     if (!soft) {
-      var draw_me = prepare_data_to_graph();
+      self.draw_me = prepare_data_to_graph();
 
-      network_layout.nodes(draw_me.all).links(draw_me.edges);
-      update_network_string(draw_me.nodes.length, draw_me.edges.length);
+      network_layout.nodes(self.draw_me.all).links(self.draw_me.edges);
+      update_network_string(self.draw_me.nodes.length, self.draw_me.edges.length);
 
       var edge_set = {};
 
-      _.each(draw_me.edges, (d) => {
+      _.each(self.draw_me.edges, (d) => {
         d.pull = 0.0;
         var tag;
 
@@ -6856,7 +6026,7 @@ var hivtrace_cluster_network_graph = function (
 
       link = self.network_svg
         .selectAll(".link")
-        .data(draw_me.edges, (d) => d.id);
+        .data(self.draw_me.edges, (d) => d.id);
 
       //link.enter().append("line").classed("link", true);
       link.enter().append("path").classed("link", true);
@@ -6883,7 +6053,7 @@ var hivtrace_cluster_network_graph = function (
 
       rendered_nodes = self.network_svg
         .selectAll(".node")
-        .data(draw_me.nodes, (d) => d.id);
+        .data(self.draw_me.nodes, (d) => d.id);
 
       rendered_nodes.exit().remove();
 
@@ -6894,7 +6064,7 @@ var hivtrace_cluster_network_graph = function (
       rendered_nodes.enter().append("g").append("path");
 
       rendered_clusters = self.network_svg.selectAll(".cluster-group").data(
-        draw_me.clusters.map((d) => d),
+        self.draw_me.clusters.map((d) => d),
         (d) => d.cluster_id
       );
 
@@ -6909,7 +6079,8 @@ var hivtrace_cluster_network_graph = function (
         .on("mouseout", cluster_pop_off)
         .call(network_layout.drag().on("dragstart", cluster_pop_off));
 
-      self.draw_cluster_table(
+      tables.draw_cluster_table(
+        self,
         self.extra_cluster_table_columns,
         self.cluster_table
       );
@@ -7028,7 +6199,8 @@ var hivtrace_cluster_network_graph = function (
 
         // draw subcluster tables
 
-        self.draw_cluster_table(
+        tables.draw_cluster_table(
+          self, 
           self.extra_subcluster_table_columns,
           self.subcluster_table,
           {
@@ -7043,9 +6215,9 @@ var hivtrace_cluster_network_graph = function (
         );
       }
       if (self._is_CDC_) {
-        self.draw_extended_node_table();
+        tables.draw_extended_node_table(self);
       } else {
-        self.draw_node_table(self.extra_node_table_columns);
+        tables.draw_node_table(self, self.extra_node_table_columns);
       }
     } else {
       rendered_nodes = self.network_svg.selectAll(".node");
@@ -7365,7 +6537,7 @@ var hivtrace_cluster_network_graph = function (
     }
   }
 
-  function collapse_cluster(x, keep_in_q) {
+  self.collapse_cluster = function (x, keep_in_q) {
     self.needs_an_update = true;
     x.collapsed = true;
     currently_displayed_objects -= self.cluster_sizes[x.cluster_id - 1] - 1;
@@ -7379,7 +6551,7 @@ var hivtrace_cluster_network_graph = function (
     return x.children.length;
   }
 
-  function expand_cluster(x, copy_coord) {
+  self.expand_cluster = function (x, copy_coord) {
     self.needs_an_update = true;
     x.collapsed = false;
     currently_displayed_objects += self.cluster_sizes[x.cluster_id - 1] - 1;
@@ -7574,7 +6746,7 @@ var hivtrace_cluster_network_graph = function (
   }
 
   function attribute_pairwise_distribution(id, dim, the_map, only_expanded) {
-    var scan_from = only_expanded ? draw_me.edges : self.edges;
+    var scan_from = only_expanded ? self.draw_me.edges : self.edges;
     var the_matrix = [];
     for (var i = 0; i < dim; i += 1) {
       the_matrix.push([]);
@@ -7854,7 +7026,7 @@ var hivtrace_cluster_network_graph = function (
             var cluster =
               self.clusters[self.cluster_mapping[open_cluster_queue[k]]];
             leftover -= cluster.children.length - 1;
-            collapse_cluster(cluster, true);
+            self.collapse_cluster(cluster, true);
           }
           if (k || open_cluster_queue.length) {
             open_cluster_queue.splice(0, k);
@@ -7862,7 +7034,7 @@ var hivtrace_cluster_network_graph = function (
         }
 
         if (leftover <= 0) {
-          expand_cluster(d, !move_out);
+          self.expand_cluster(d, !move_out);
         }
       }
 
@@ -7904,7 +7076,7 @@ var hivtrace_cluster_network_graph = function (
   }
 
   function collapse_cluster_handler(d, do_update) {
-    collapse_cluster(self.clusters[self.cluster_mapping[d.cluster]]);
+    self.collapse_cluster(self.clusters[self.cluster_mapping[d.cluster]]);
     if (do_update) {
       self.update(false, 0.4);
     }
@@ -7967,7 +7139,7 @@ var hivtrace_cluster_network_graph = function (
   self.collapse_some_clusters = function (subset) {
     subset = subset || self.clusters;
     subset.forEach((x) => {
-      if (!x.collapsed) collapse_cluster(x);
+      if (!x.collapsed) self.collapse_cluster(x);
     });
     self.update();
   };
@@ -8403,9 +7575,9 @@ var hivtrace_cluster_network_graph = function (
 
       self.update_clusters_with_injected_nodes(null, null, annotation);
       if (self._is_CDC_) {
-        self.draw_extended_node_table(self.json.Nodes);
+        tables.draw_extended_node_table(self, self.json.Nodes);
       } else {
-        self.draw_node_table(self.extra_node_table_columns, self.json.Nodes);
+        tables.draw_node_table(self, self.extra_node_table_columns, self.json.Nodes);
       }
       if (!self.extra_cluster_table_columns) {
         self.extra_cluster_table_columns = [];
@@ -8601,7 +7773,7 @@ var hivtrace_cluster_network_graph = function (
                 cluster.cluster_id,
               ],
 
-              callback: function (element, payload) {
+              callback: function (self, element, payload) {
                 var this_cell = d3.select(element);
                 this_cell.text(Number(payload[0]) + " " + annotation + " nodes. ");
                 var other_clusters = [];
@@ -8707,11 +7879,14 @@ var hivtrace_cluster_network_graph = function (
       if (self.extra_cluster_table_columns) {
         self.extra_cluster_table_columns =
           self.extra_cluster_table_columns.concat(injected_column);
+
+          console.log("HIT")
       } else {
         self.extra_cluster_table_columns = injected_column;
       }
 
-      self.draw_cluster_table(
+      tables.draw_cluster_table(
+        self, 
         self.extra_cluster_table_columns,
         self.cluster_table
       );
@@ -8725,7 +7900,8 @@ var hivtrace_cluster_network_graph = function (
         } else {
           self.extra_subcluster_table_columns = injected_column_subcluster;
         }
-        self.draw_cluster_table(
+        tables.draw_cluster_table(
+          self,
           self.extra_subcluster_table_columns,
           self.subcluster_table,
           { subclusters: true, "no-clusters": true }
