+++ ui-tests/clusterOI.spec.js
@@ -6,9 +6,11 @@ test.beforeEach(async ({ page }) => {
   errors = [];
   page.on('console', msg => {
     if (msg.type() === 'error') {
-      console.log(msg.text());
       errors.push(msg.text());
     }
+  })
+  page.on("pageerror", (err) => {
+    errors.push(err.message);
   })
 
   await page.goto('http://127.0.0.1:8080/html/priority-sets-args.html?network=../ui-tests/data/network.json');
